# packages/contracts/openapi/cirda-v1.yaml
openapi: 3.1.0
info:
  title: CIRDA Control Plane API
  version: 0.1.0
paths:
...
  /api/v1/edges:
    get:
      tags:
      - edges
      summary: List Edges
      operationId: list_edges_api_v1_edges_get
      parameters:
      - name: layer
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Layer
      - name: source_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Id
      - name: target_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Target Id
      - name: as_of
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/edges/necessity-suggestions:
    get:
      tags:
      - edges
      summary: Necessity Suggestions
      operationId: necessity_suggestions_api_v1_edges_necessity_suggestions_get
      parameters:
      - name: source_id
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Source Id
      - name: as_of
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NecessityHintListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/edges/{edge_id}:
    get:
      tags:
      - edges
      summary: Get Edge
      operationId: get_edge_api_v1_edges__edge_id__get
      parameters:
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          title: Edge Id
      - name: as_of
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - edges
      summary: Patch Edge Necessity
      operationId: patch_edge_necessity_api_v1_edges__edge_id__patch
      parameters:
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          title: Edge Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeNecessityUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdgeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/edges/{edge_id}/evidence:
    get:
      tags:
      - edges
      summary: Edge Evidence
      operationId: edge_evidence_api_v1_edges__edge_id__evidence_get
      parameters:
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          title: Edge Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Edge Evidence Api V1 Edges  Edge Id  Evidence Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/decisions/evaluate:
    post:
      tags:
      - decisions
      summary: Evaluate Decision
      operationId: evaluate_decision_api_v1_decisions_evaluate_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionEvaluateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/probes/apply:
    post:
      tags:
      - decisions
      summary: Apply Probes
      operationId: apply_probes_api_v1_decisions_probes_apply_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProbeApplyRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProbeApplyResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/by-entity/{entity_id}:
    get:
      tags:
      - decisions
      summary: List Entity Decisions
      operationId: list_entity_decisions_api_v1_decisions_by_entity__entity_id__get
      parameters:
      - name: entity_id
        in: path
        required: true
        schema:
          type: string
          title: Entity Id
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}:
    get:
      tags:
      - decisions
      summary: Get Decision
      operationId: get_decision_api_v1_decisions__decision_id__get
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/decisions/{decision_id}/rerun:
    post:
      tags:
      - decisions
      summary: Rerun Decision
      operationId: rerun_decision_api_v1_decisions__decision_id__rerun_post
      parameters:
      - name: decision_id
        in: path
        required: true
        schema:
          type: string
          title: Decision Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionReport'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DecisionEvaluateRequest:
      properties:
        entity_id:
          type: string
          title: Entity Id
        as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
        change_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Type
      additionalProperties: false
      type: object
      required:
      - entity_id
      title: DecisionEvaluateRequest
    DecisionListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DecisionReport'
          type: array
          title: Items
        page:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - page
      title: DecisionListResponse
    DecisionRationale:
      properties:
        summary:
          type: string
          title: Summary
        details:
          items:
            type: string
          type: array
          title: Details
        reason_codes:
          items:
            type: string
          type: array
          title: Reason Codes
        blast_radius:
          additionalProperties: true
          type: object
          title: Blast Radius
        coverage_breakdown:
          additionalProperties: true
          type: object
          title: Coverage Breakdown
        suggested_runbook:
          items:
            $ref: '#/components/schemas/RunbookActionSchema'
          type: array
          title: Suggested Runbook
        suggested_probes:
          items:
            $ref: '#/components/schemas/ProbePlanSchema'
          type: array
          title: Suggested Probes
      additionalProperties: false
      type: object
      required:
      - summary
      title: DecisionRationale
    DecisionReport:
      properties:
        decision_id:
          type: string
          title: Decision Id
        entity_id:
          type: string
          title: Entity Id
        verdict:
          $ref: '#/components/schemas/VerdictSchema'
        coverage:
          type: number
          title: Coverage
        truncated:
          type: boolean
          title: Truncated
          default: false
        reason_codes:
          items:
            type: string
          type: array
          title: Reason Codes
        rationale:
          anyOf:
          - $ref: '#/components/schemas/DecisionRationale'
          - additionalProperties: true
            type: object
          title: Rationale
        limitations_footer:
          type: string
          title: Limitations Footer
        as_of:
          type: string
          format: date-time
          title: As Of
        engine_version:
          type: string
          title: Engine Version
        change_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Change Type
        supersedes:
          anyOf:
          - type: string
          - type: 'null'
          title: Supersedes
        superseded_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Superseded By
        created_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Created By
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        paths:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Paths
      additionalProperties: false
      type: object
      required:
      - decision_id
      - entity_id
      - verdict
      - coverage
      - rationale
      - limitations_footer
      - as_of
      - engine_version
      title: DecisionReport
    EdgeListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EdgeResponse'
          type: array
          title: Items
        page:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - page
      title: EdgeListResponse
    EdgeNecessityUpdate:
      properties:
        necessity:
          type: string
          title: Necessity
          description: 'Operator necessity annotation: required|optional|redundant|fallback|unknown'
      type: object
      required:
      - necessity
      title: EdgeNecessityUpdate
    EdgeResponse:
      properties:
        edge_id:
          type: string
          title: Edge Id
        source_id:
          type: string
          title: Source Id
        target_id:
          type: string
          title: Target Id
        relation:
          type: string
          title: Relation
        layer:
          type: string
          title: Layer
        confidence:
          type: number
          title: Confidence
        necessity:
          type: string
          title: Necessity
          default: unknown
        evidence_count:
          type: integer
          title: Evidence Count
          default: 0
      type: object
      required:
      - edge_id
      - source_id
      - target_id
      - relation
      - layer
      - confidence
      title: EdgeResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NecessityHintListResponse:
      properties:
        source_id:
          type: string
          title: Source Id
        items:
          items:
            $ref: '#/components/schemas/NecessityHintSchema'
          type: array
          title: Items
        mode:
          type: string
          title: Mode
          default: suggest_only
      type: object
      required:
      - source_id
      - items
      title: NecessityHintListResponse
    NecessityHintSchema:
      properties:
        edge_id:
          type: string
          title: Edge Id
        source_id:
          type: string
          title: Source Id
        target_id:
          type: string
          title: Target Id
        current_necessity:
          type: string
          title: Current Necessity
        suggested_necessity:
          type: string
          title: Suggested Necessity
        confidence:
          type: number
          title: Confidence
        rationale:
          type: string
          title: Rationale
        signals:
          additionalProperties: true
          type: object
          title: Signals
      type: object
      required:
      - edge_id
      - source_id
      - target_id
      - current_necessity
      - suggested_necessity
      - confidence
      - rationale
      title: NecessityHintSchema
    PageMeta:
      properties:
        total:
          type: integer
          title: Total
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
      additionalProperties: false
      type: object
      required:
      - total
      - offset
      - limit
      title: PageMeta
    ProbeApplyRequest:
      properties:
        entity_id:
          type: string
          title: Entity Id
        channels:
          items:
            type: string
          type: array
          minItems: 1
          title: Channels
        as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      additionalProperties: false
      type: object
      required:
      - entity_id
      - channels
      title: ProbeApplyRequest
    ProbeApplyResult:
      properties:
        entity_id:
          type: string
          title: Entity Id
        channels:
          items:
            type: string
          type: array
          title: Channels
        mode:
          type: string
          title: Mode
          default: suppression_lift
        coverage_before:
          type: number
          title: Coverage Before
        coverage_after:
          type: number
          title: Coverage After
        expected_delta_c:
          type: number
          title: Expected Delta C
        actual_delta_c:
          type: number
          title: Actual Delta C
        suppressed_channels_after:
          items:
            type: string
          type: array
          title: Suppressed Channels After
        as_of:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: As Of
      additionalProperties: false
      type: object
      required:
      - entity_id
      - channels
      - coverage_before
      - coverage_after
      - expected_delta_c
      - actual_delta_c
      title: ProbeApplyResult
    ProbePlanSchema:
      properties:
        entity_id:
          type: string
          title: Entity Id
        channels:
          items:
            type: string
          type: array
          title: Channels
        rationale:
          type: string
          title: Rationale
        expected_delta_c:
          type: number
          title: Expected Delta C
          default: 0.0
      additionalProperties: false
      type: object
      required:
      - entity_id
      - rationale
      title: ProbePlanSchema
...
    RunbookActionSchema:
      properties:
        stage:
          type: string
          title: Stage
        description:
          type: string
          title: Description
        entity_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Entity Id
      additionalProperties: false
      type: object
      required:
      - stage
      - description
      title: RunbookActionSchema
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerdictSchema:
      type: string
      enum:
      - UNSAFE
      - SAFE
      - INDETERMINATE
      title: VerdictSchema